Feat: added blogs and updates to product page
