+++ src/pages/products/closed-messaging-checklist.mdx
@@ -1,4 +1,7 @@
-# Responsible Data Collection and Sharing from Closed Messaging Apps
+import { LatestProductBlogsUpdates } from "../../components/LatestProductBlogsUpdates"
+
+<h1 style={{lineHeight:"1.5em"}}>Responsible Data Collection and Sharing from Closed Messaging Apps</h1>
+
 ## A Checklist 
 (*Last Updated*: April 29, 2021)
 
@@ -98,3 +101,5 @@ By codifying responsible data management practices prior to the study, researche
 
 
 ---
+{/* Add projects here to show the blogs and updates */}
+<LatestProductBlogsUpdates projects={[]}/>
