+++ src/pages/products/dau.mdx
@@ -1,5 +1,6 @@
 import { Anchor, Text, Box } from "grommet"
 import { PlainLink } from "../../components/atomic/TattleLinks"
+import { LatestProductBlogsUpdates } from "../../components/LatestProductBlogsUpdates"
 
 # Deepfakes Analysis Unit
 
@@ -28,3 +29,5 @@ The [DAU's](https://dau.mcaindia.in/) (Deepfakes Analysis Unit) mission is to ge
 <a href="https://mcaindia.in/"><Text size={"small"}>Misinformation Combat Alliance</Text></a>
 <Text size={"small"}>Tattle</Text>
 <Text size={"small"}>Meta (Funding Partner)</Text>
+
+<LatestProductBlogsUpdates projects={["dau"]}/>
