+++ src/pages/products/gftw.mdx
@@ -1,3 +1,5 @@
+import { LatestProductBlogsUpdates } from "../../components/LatestProductBlogsUpdates"
+
 ## Introduction to the Study
 
 Any reconceptualization of the Internet requires rethinking of incentives not only for content creation but also for content discovery and 
@@ -50,9 +52,12 @@ We aim to continue to do this kind of research on incentives and content creatio
 * [Preprint](https://psyarxiv.com/nykxz)
 * [Final report for the project](https://community.interledger.org/denntenna/final-report-for-using-web-monetisation-for-incentivising-sharing-of-good-content-1gef)
 
+
 ## Key Team Members
 
 * Hansika Kapoor, Monk Prayogshala
 * Denny George, Tattle
 * Arathy Puthillam, Monk Prayogshala
 * Yash Budhwar, Tattle
+
+<LatestProductBlogsUpdates projects={["Grant for the Web"]}/>
