+++ src/pages/products/github-indices.mdx
@@ -1,4 +1,6 @@
-# Project with GitHub to develop indices for public policy, international development and economic development. 
+import { LatestProductBlogsUpdates } from "../../components/LatestProductBlogsUpdates"
+
+<h1 style={{lineHeight:"1.5em"}}> Project with GitHub to develop indices for public policy, international development and economic development.</h1> 
 
 This project has concluded. 
 
@@ -20,3 +22,5 @@ Between February 2022 to August 2022 we conduced one-on-one interviews and focus
 ## Outcomes:
 * A report by GitHub is forthcoming. 
 * The work was presented at RightsCon 2023. The session was titled: 'Developing indexes on internet platform data for social sector research.' 
+
+<LatestProductBlogsUpdates projects={["github indices"]} />
