+++ src/pages/products/jod-bot.mdx
@@ -1,3 +1,5 @@
+import { LatestProductBlogsUpdates } from "../../components/LatestProductBlogsUpdates"
+
 # Jod Bot
 
 Jod Bot is a Telegram Bot that lets you push content to Tattle's archive and add tags to it. 
@@ -5,3 +7,5 @@ Jod Bot is a Telegram Bot that lets you push content to Tattle's archive and add
 [Go to Service](https://t.me/TattleJodBot) (Telegram Bot Link)
 
 [Got To Code Repository](https://github.com/tattle-made/archive-telegram-bot)
+
+<LatestProductBlogsUpdates projects={["jod bot"]}/>
