+++ src/pages/products/khoj.mdx
@@ -1,4 +1,5 @@
 import {Box, Text} from 'grommet'
+import { LatestProductBlogsUpdates } from "../../components/LatestProductBlogsUpdates"
 
 # Khoj
 
@@ -14,7 +15,7 @@ Khoj is a service that makes it easy to verify messages you receive on your soci
 
 ![Khoj App and Community UI](../../images/khoj-app-community-ui.png)
 
-<!-- [Download our app from the Play Store](https://play.google.com/store/apps/details?id=in.co.tattle.khoj) -->
+{/* [Download our app from the Play Store](https://play.google.com/store/apps/details?id=in.co.tattle.khoj) */}
 
 If you wish to answer queries regarding social media posts from users using our Community UI, please reach out to us at admin@tattle.co.in
 
@@ -24,6 +25,8 @@ Tattle Khoj is the front end for Tattle's search API, but plugged only to the da
 
 ![Gif showing Khoj being used](https://tattle-media.s3.amazonaws.com/khoj-demo.gif)
 
-<!-- [Go to Service](https://services.tattle.co.in/khoj/) -->
+{/* <!-- [Go to Service](https://services.tattle.co.in/khoj/) --> */}
 
 [Go to Code Repository](https://github.com/tattle-made/kosh)
+
+<LatestProductBlogsUpdates projects={["khoj"]}/>
