+++ src/pages/products/ogbv.jsx
@@ -3,6 +3,7 @@ import NarrowSection from "../../components/atomic/layout/narrow-section"
 import DefaultLayout from "../../components/default-layout"
 import NarrowContentWrapper from "../../components/atomic/layout/narrow-content-wrapper"
 import { Anchor, Box, Heading, Image, List, Paragraph, Text } from "grommet"
+import { LatestProductBlogsUpdates } from "../../components/LatestProductBlogsUpdates"
 
 const ogbv = () => (
   <DefaultLayout>
@@ -146,6 +147,10 @@ const ogbv = () => (
           )}
         </List>
       </NarrowSection>
+
+      <NarrowSection><LatestProductBlogsUpdates projects={["uli","ogbv"]} /></NarrowSection>
+
+
       <NarrowSection>
         <Heading level={3}>Current Team</Heading>
         <Text size="small">
