+++ src/pages/products/whatsapp-archiver.mdx
@@ -1,3 +1,5 @@
+import { LatestProductBlogsUpdates } from "../../components/LatestProductBlogsUpdates"
+
 # Whatsapp Archiver
 
 ## About the Archiver
@@ -45,3 +47,5 @@ Tattle is primarily interested in the content itself, and not the people or netw
 
 ## WhatsApp Scraper:
 [Go to Service/Code Repository](https://github.com/tattle-made/whatsapp-scraper/tree/master/python_scraper)
+
+<LatestProductBlogsUpdates projects={["whatsapp archiver"]}/>
